Add async tests
