+++ rust/geoparquet/src/test/geoarrow_data/mod.rs
@@ -1,4 +1,2 @@
 mod example;
-#[cfg(feature = "async")]
-mod example_async;
 mod example_crs;
